Extract aggregate accumulator into its own module
Move the AggrValue accumulator and its NULL handling out of the GROUP BY
state module so the same accumulation logic can be reused by other callers.
No behavior change.
