+++ core/src/executor/aggregate.rs
@@ -1,3 +1,4 @@
+pub(crate) mod accumulator;
 mod state;
 
 use {
